// src/app/features/home/chat/Models/filter-system-user-dto.ts
export interface FilterSystemUserDTO {
  userGroupId?: number;
  roleId?: number;
  levelId?: number;
  govenmentId?: number;
  healthAdministrationId?: number;
  incidentSourceId?: number;
  positionId?: number;
  departmentId?: number;
  fullName?: string;
  phoneNo?: string;
  email?: string;
  address?: string;
  userTypeI?: number;
  externalLabId?: number;
  organizationId?: number;
  pageSize?: number;
  pageIndex?: number;
}