// src/app/models/chat-filter.ts
export interface ChatFilter {
  currentUserLevelID?: number;
  fullName?: string;
  roleId?: number[];
  levelId?: number[];
  govenmentId?: number[];
  healthAdministrationId?: number[];
  incidentSourceId?: number[];
  positionId?: number[];
  departmentId?: number[];
  organizationId?: number[];
  incidentSourceTypeId?: number[];

}